Initial port of googletest to Unikraft
This is our initial port of googletest to Unikraft as an external
library. libcxx is required. This library should be added last to the
dependency list, e.g.,:

    ...:$(UK_LIBS)/libunwind:$(UK_LIBS)/compiler-rt:$(UK_LIBS)/libcxxabi:
        $(UK_LIBS)/libcxx:$(UK_LIBS)/newlib:$(UK_LIBS)/googletest:...

This library has a config option called LIBGOOGLETEST_BUILD_MAIN that builds the
gtest_main.cc file; this is because some libraries do not have any main function
in their unit tests and expect googletest to provide it.

Signed-off-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@stud.acs.upb.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>